requirements.txt
python-dotenv==1.0.0
pymongo==4.7.3
dnspython==2.6.1
certifi==2024.2.2
requests==2.32.3
beautifulsoup4==4.12.3
schedule==1.2.1
pytz==2024.1
flask==3.0.3
flask-wtf==1.2.1
flask-bootstrap4==4.1.1
